+++ CHANGELOG.md
@@ -5,8 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.28] - 2026-05-27
+
 - Fixed `ValidateReportOrder` entity: removed non-existing field "status"
- 
+
 ## [0.11.27] - 2026-05-21
 
 - Added `ReportsClient` with `validateReport()` method to validate reports via `PATCH /reports/{reportReference}`
@@ -673,7 +675,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Entity Price: use PriceBand now, which is in the Catalog namespace where it belongs
 - Entity Service: the root entity is deprecated, and the one in the Catalog namespace should be used
 
-[Unreleased]: https://github.com/ArrowSphere/public-api-client/compare/0.11.27...HEAD
+[Unreleased]: https://github.com/ArrowSphere/public-api-client/compare/0.11.28...HEAD
+[0.11.28]: https://github.com/ArrowSphere/public-api-client/compare/0.11.27...0.11.28
 [0.11.27]: https://github.com/ArrowSphere/public-api-client/compare/0.11.26...0.11.27
 [0.11.26]: https://github.com/ArrowSphere/public-api-client/compare/0.11.25...0.11.26
 [0.11.25]: https://github.com/ArrowSphere/public-api-client/compare/0.11.24...0.11.25